requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.20
python-docx==1.1.0
PyPDF2==3.0.1
sentence-transformers==2.3.0
faiss-cpu==1.12.0
requests==2.31.0
stripe==7.8.0
pandas==2.1.4
openpyxl==3.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
aiofiles==23.2.1
pydantic-settings==2.2.1
httpx==0.27.0
numpy==1.26.2
torch==2.2.2
transformers==4.36.2
scikit-learn==1.3.2
pydantic==2.6.4
huggingface-hub==0.19.4
jwt
